requirements.txt
black==23.1.0
click==8.1.3
greenlet==2.0.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
py4j==0.10.9.5
pyspark==3.3.1
python-dotenv==0.21.1
SQLAlchemy==2.0.0
tomli==2.0.1
typing-extensions==4.4.0
